Remove Timetable class
